エラー周りの機能を実装
・manifest.jsonファイルがない時のエラー
・自動バージョンアップデート数未入力時のエラー
・インストールされていないものを表示する警告ダイアログ
・正しく実行されなかった際に自動バージョンアップの更新をなかったことにする等の機能を実装。
