+++ WindowsFormsApp2/JsonObjectConfig.cs
@@ -2,16 +2,20 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
 namespace WindowsFormsApp2
 {
     public class JsonObjectConfig
     {
-        
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
         public string html { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
         public List<string> js { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
         public List<string> css { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
         public List<string> required_params { get; set; }
     }
 }
