+++ WindowsFormsApp2/WindowsFormsApp2.csproj
@@ -81,23 +81,25 @@
   <ItemGroup>
     <Compile Include="ComandProcess.cs" />
     <Compile Include="Controller.cs" />
-    <Compile Include="Form1.cs">
+    <Compile Include="ExecutionCondition.cs" />
+    <Compile Include="MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Form1.Designer.cs">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <Compile Include="MainForm.Designer.cs">
+      <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Form2.cs">
+    <Compile Include="PluginNameSettingForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Form2.Designer.cs">
-      <DependentUpon>Form2.cs</DependentUpon>
+    <Compile Include="PluginNameSettingForm.Designer.cs">
+      <DependentUpon>PluginNameSettingForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Form3.cs">
+    <Compile Include="ManifestVisibleForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Form3.Designer.cs">
-      <DependentUpon>Form3.cs</DependentUpon>
+    <Compile Include=".Designer.cs">
+      <DependentUpon>ManifestVisibleForm.cs</DependentUpon>
+      <SubType>Form</SubType>
     </Compile>
     <Compile Include="FormData.cs" />
     <Compile Include="JsonObjectConfig.cs" />
@@ -106,14 +108,14 @@
     <Compile Include="ManifestJsonData.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <EmbeddedResource Include="Form1.resx">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <EmbeddedResource Include="MainForm.resx">
+      <DependentUpon>MainForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Form2.resx">
-      <DependentUpon>Form2.cs</DependentUpon>
+    <EmbeddedResource Include="PluginNameSettingForm.resx">
+      <DependentUpon>PluginNameSettingForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Form3.resx">
-      <DependentUpon>Form3.cs</DependentUpon>
+    <EmbeddedResource Include="ManifestVisibleForm.resx">
+      <DependentUpon>ManifestVisibleForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
