+++ WindowsFormsApp2/echo.bat
@@ -1,2 +1,3 @@
 cd %1
 kintone-plugin-uploader --watch --base-url %2 --username %3 --password %4 plugin.zip
+exit %errorlevel%
