+++ WindowsFormsApp2/echo3.bat
@@ -1,0 +1,2 @@
+npm -v
+exit %errorlevel%
