+++ WindowsFormsApp2/echo4.bat
@@ -1,0 +1,2 @@
+kintone-plugin-packer
+exit %errorlevel%
